feat(deploy): :clown_face: mock the `production` deployment environment
The added `deploy/production@host` folder mocks the 'production' deployment environment. This folder was created as a "hardlink copy" (in the sense of the `ln` PowerShell command) of the `new-release@rel` folder, which is:  - The folder are standard folders. - The files are hardlinks pointing to the same contents as the "original" files.  Note that a CAUTION tag to warn about the presence of hardlinks was added to every concerned file.