style(abstracts): :pencil2: update abstracts to pure markdown syntax
In early versions, using html in abstracts' bodies was required due to the use of surrounding html tags. This turns out to not be needed anymore.

style(abstracts): :pencil2: refine assets abstracts to  differentiate between the user's and the templates'

In addition to the markdown fix, abstracts were modified to better show the difference between the user's assets found in the `resource` folder and the functional ones found in the `templates` folder.

Also, a redundancy within the `templates` folder has been fixed: out of the 2 `assets` models populating `templates`, only this within the `directories` subfolder was kept.